requirements.txt
# Core dependencies
streamlit>=1.28.1
openai>=1.3.5
python-dotenv>=1.0.0
pydantic>=2.5.0

# Document processing
PyPDF2>=3.0.1
python-docx>=0.8.11
pymupdf>=1.23.14
pillow>=10.1.0
pytesseract>=0.3.10

# Vector database and embeddings
chromadb>=0.4.18

# Data handling
pandas>=2.1.3
numpy>=1.24.3

# Production server
gunicorn>=21.2.0
